fix: add every event to its component
